Configure the time zone to EST.
This is the time zone where Launch Academy is physically located
and most of our users reside. Although timestamps are stored in GMT
in the database they will be converted to EST in the view.

Eventually the user should be able to configure their preferred time
zone rather than forcing a site-wide setting.